예적금 api 완료

+++ src/main/java/com/heeha/domain/signDeposit/entity/SignDeposit.java
@@ -34,7 +34,7 @@ public class SignDeposit extends BaseEntity {
     private InstallmentMethodType installmentMethod; // 적립 방법
     private MaturityClassType maturityClass; // 만기해지구분
     private AutoCancelType autoCancel; // 자동 해지 시점
-    private Integer monthlyAmount; // 월 납입 금액
+    private Long monthlyAmount; // 월 납입 금액
 
     private Integer contractYears; // 계약 햇수 setSubscriptionPeriod
     private Boolean snsNotice; // SNS 만기 알림 setSmsNotification
